requirements/requirements.txt
amqp==2.5.2
ansible==2.9.24
asn1crypto==0.24.0
bcrypt==3.1.4
billiard==3.6.3.0
boto3==1.18.11
botocore==1.21.11
celery==4.4.2
certifi==2018.1.18
cffi==1.13.2
chardet==3.0.4
configparser==3.5.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.2
decorator==4.1.2
Django==3.1.13
django-auth-ldap==2.2.0
django-bootstrap3==14.2.0
django-celery-beat==2.0
django-filter==2.4.0
django-formtools==2.2
django-ranged-response==0.2.0
django-redis-cache==2.1.1
django-rest-swagger==2.2.0
django-simple-captcha==0.5.13
django-timezone-field==4.0
djangorestframework==3.12.2
djangorestframework-bulk==0.2.1
docutils==0.14
ecdsa==0.13.3
enum-compat==0.0.2
ephem==3.7.6.0
eventlet==0.31.1
future==0.16.0
ForgeryPy3==0.3.1
greenlet==0.4.14
gunicorn==19.9.0
idna==2.6
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.11.3
jmespath==0.9.3
kombu==4.6.8
ldap3==2.4
MarkupSafe==1.1.1
mysqlclient==2.0.1
olefile==0.44
openapi-codec==1.3.2
paramiko==2.7.2
passlib==1.7.1
Pillow==8.3.2
pyasn1==0.4.8
pycparser==2.19
pycryptodome==3.10.1
pycryptodomex==3.10.1
pyotp==2.2.6
PyNaCl==1.2.1
python-dateutil==2.6.1
#python-gssapi==0.6.4
pytz==2018.3
PyYAML==5.4
redis==3.5.3
requests==2.25.1
jms-storage==0.0.40
s3transfer==0.5.0
simplejson==3.13.2
six==1.11.0
sshpubkeys==3.1.0
uritemplate==3.0.0
urllib3==1.26.5
vine==1.3.0
drf-yasg==1.20.0
Werkzeug==0.15.3
drf-nested-routers==0.91
aliyun-python-sdk-core-v3==2.9.1
aliyun-python-sdk-ecs==4.10.1
rest_condition==1.0.3
python-ldap==3.3.1
tencentcloud-sdk-python==3.0.477
django-radius==1.4.0
django-redis-sessions==0.6.1
unicodecsv==0.14.1
python-daemon==2.2.3
httpsig==1.3.0
treelib==1.5.3
django-proxy==1.2.1
flower==0.9.3
channels-redis==3.2.0
channels==2.4.0
daphne==2.4.1
psutil==5.6.6
django-cas-ng==4.0.1
python-cas==1.5.0
ipython
huaweicloud-sdk-python==1.0.21
django-redis==4.11.0
python-redis-lock==3.7.0
jumpserver-django-oidc-rp==0.3.7.8
django-mysql==3.9.0
gmssl==3.2.1
azure-mgmt-compute==4.6.2
azure-mgmt-network==2.7.0
msrestazure==0.6.4
adal==1.2.5
openpyxl==3.0.5
pyexcel==0.6.6
pyexcel-xlsx==0.6.0
data-tree==0.0.1
pyvmomi==7.0.1
termcolor==1.1.0
azure-identity==1.5.0
azure-mgmt-subscription==1.0.0
qingcloud-sdk==1.2.12
django-simple-history==3.0.0
google-cloud-compute==0.5.0
PyMySQL==1.0.2
cx-Oracle==8.2.1
psycopg2-binary==2.9.1
alibabacloud_dysmsapi20170525==2.0.2
geoip2==4.4.0
html2text==2020.1.16
python-novaclient==11.0.1
pyzipper==0.3.5
python3-saml==1.12.0
python-keystoneclient==4.3.0
pymssql==6.0
kubernetes==21.7.0
websocket-client==1.2.3
